pyproject.toml
[tool.poetry]
name = "pb-buddy"
version = "0.1.0"
description = "A package to help you ride better."
authors = ["Dustin Andrews"]
license = "*"

[tool.poetry.dependencies]
python = "~3.11"
numpy = "<1.29"
pandas = "^2.2.3"
beautifulsoup4 = "^4.12.3"
requests = "^2.32.3"
tqdm = "^4.66.5"
joblib = "^1.4.2"
python-dotenv = "^1.0.1"
certifi = "^2024.8.30"
pymongo = "^4.10.1"
matplotlib = "^3.9.2"
jinja2 = "^3.1.4"
jupytext = "^1.16.4"
nbconvert = "^7.16.4"
azure-storage-blob = "^12.23.1"
seaborn = "^0.13.2"
lxml = "^5.3.0"
html5lib = "^1.1"
pandas-datareader = "^0.10.0"
black = "^24.10.0"
fire = "^0.7.0"
playwright = "^1.47.0"
rapidfuzz = "^3.10.0"
scikit-learn = "<1.5.2"
tenacity = "^9.0.0"
openai = "^1.51.2"
duckdb = "^1.1.1"
s3fs = "^2024.9.0"
protobuf = "^5.28.2"
pydantic = "^2.9.2"
pyarrow = "^18.1.0"
yfinance = "^0.2.43"


[tool.poetry.dev-dependencies]
ipykernel = "*"

[tool.poetry.group.dev.dependencies]
pytest = "*"

[tool.poetry.group.modelling]
optional = true

[tool.poetry.group.modelling.dependencies]
autogluon-multimodal = "^1.1.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
